+++ src/utils/index.ts
@@ -3,3 +3,5 @@ export * from './address';
 export * from './token';
 export * from './age';
 export * from './compressImage';
+export * from './getCoordsFromAddress';
+export * from './getEmptyProfile';
